♻️ Refactor popup list to context
