+++ src/pages/index.tsx
@@ -1,36 +1,19 @@
-import React, { FC, useCallback } from "react";
-import { useRecoilState, useRecoilValue } from "recoil";
+import React, { FC } from "react";
 import styled from "styled-components";
 
 import NaverMap from "components/NaverMap";
-import Popup from "components/Popup";
+import PopupContext from "components/PopupContext";
 import RightPopup from "components/RightPopup";
 import Sidebar from "components/Sidebar";
-import popupState from "utils/states/popupState";
 
-const IndexPage: FC = () => {
-  const [popupList, setPopupList] = useRecoilState(popupState);
-
-  const handlePopupRemove = useCallback(
-    (index: number) => {
-      const newList = [...popupList];
-      newList.splice(index, 1);
-      setPopupList(newList);
-    },
-    [popupList],
-  );
-
-  return (
-    <PageWrapper>
-      <Sidebar />
-      <NaverMap />
-      <RightPopup />
-      {popupList.map((p, index) => (
-        <Popup key={index} onClose={() => handlePopupRemove(index)} popup={p} />
-      ))}
-    </PageWrapper>
-  );
-};
+const IndexPage: FC = () => (
+  <PageWrapper>
+    <Sidebar />
+    <NaverMap />
+    <RightPopup />
+    <PopupContext />
+  </PageWrapper>
+);
 
 export default IndexPage;
